refactor: create reusable util for opening popup
